+++ src/widget/abstract/View.jsx
@@ -12,7 +12,9 @@ import BoardStore                           from '../../core/BoardStore';
 
 import Styles   from './View.scss';
 
-
+/**
+ * Manage drag & drop and all others actions common to all widgets in view mode
+ */
 export default class AbstractWidgetView extends Component {
 
     constructor( props ) {
@@ -44,6 +46,9 @@ export default class AbstractWidgetView extends Component {
         ];
     }
 
+    /**
+     * If widgets is not locked set edit mode
+     */
     onDoubleClick( event ) {
         event.preventDefault();
         if ( !this.props.isLockedByAnotherUser ) {
@@ -51,26 +56,30 @@ export default class AbstractWidgetView extends Component {
         }
     }
 
+    /**
+     * start drag & drop on Mouse Down if is not locked
+     */
     onMouseDown( event ) {
         if ( !this.props.isLockedByAnotherUser ) {
             this.onDragStart( event );
         }
     }
 
-    getInvertedZoom() {
-        const zoom = BoardStore.zoom;
-        return zoom >= 1 ? 1 - ( zoom - 1 ) : 1 + ( 1 - zoom );
-        //return 1;
-    }
-
+    /**
+     * Init D&D
+     * stores initial position of mouse cursor and widget
+     * @param  {event} event mouseClick event
+     */
     onDragStart( event ) {
         if ( !this.state.canDrag ) {
             return;
         }
 
-        const invertedZoom = this.getInvertedZoom();
-        this._startX = ( event.pageX * invertedZoom ) - this.props.position.x;
-        this._startY = ( event.pageY * invertedZoom ) - this.props.position.y;
+        this.initialEventX = event.pageX;
+        this.initialEventY = event.pageY;
+
+        this.initialPropsX = this.props.position.x
+        this.initialPropsY = this.props.position.y
 
         this.isDragging = true;
 
@@ -82,7 +91,13 @@ export default class AbstractWidgetView extends Component {
         this.props.actions.select();
     }
 
+    /**
+     * Manage D&D
+     * @param  {event} event New position of the cursor
+     */
     onDrag = ( event ) => {
+
+        const zoom = BoardStore.zoom;
 
         this.clearScrollTimeouts();
 
@@ -98,17 +113,17 @@ export default class AbstractWidgetView extends Component {
             this.scrollTop( true );
         }
 
-        var y = event.pageY - this._startY;
-        var x = event.pageX - this._startX;
+        //Formula to compute position of the widgets and consider the zoom
+        var x = this.initialPropsX + ((event.pageX - this.initialEventX) / zoom);
+        var y = this.initialPropsY + ((event.pageY - this.initialEventY) / zoom);
 
         x = x > 0 ? x : 0;
         y = y > 0 ? y : 0;
 
-        const inversedZoom = this.getInvertedZoom();
-
-        x = x * inversedZoom;
-        y = y * inversedZoom;
-
+        /*
+            round up values of x & y
+            example x: 478.5 y : 201 ==> x : 470 y : 200
+         */
         if ( ( Math.abs( this.props.position.x - x ) >= gridWidth ) ||
              ( Math.abs( this.props.position.y - y ) >= gridWidth ) ) {
 
@@ -119,6 +134,9 @@ export default class AbstractWidgetView extends Component {
         }
     };
 
+    /**
+     * Remove listener and unlock widget
+     */
     onDragEnd = ( event ) => {
         this.isDragging = false;
         document.removeEventListener( 'mousemove', this.onDrag );
@@ -216,6 +234,24 @@ export default class AbstractWidgetView extends Component {
             );
         }
 
+        if( this.props.isEditingByAnotherUser ){
+            return (
+                <div tabIndex="1000"
+                     style={ style }
+                     className={ className } >
+                  <div className={ Styles.isEditingByAnotherUser }>
+                        { this.props.isEditingBy } is editing...
+                  </div>
+
+                   <Menu ref="menu"
+                     position = { this.props.position }
+                     lock     = { true }
+                     display  = { true }
+                     lockName = { this.props.lockName }/>
+               </div>
+            );
+        }
+
         if( this.props.isLockedByAnotherUser ){
             return (
                 <div tabIndex="1000"
@@ -232,6 +268,7 @@ export default class AbstractWidgetView extends Component {
             );
         }
 
+
         return (
             <div tabIndex="1000"
                  style={ style }
@@ -244,7 +281,9 @@ export default class AbstractWidgetView extends Component {
                  <Menu ref="menu" menuElements={ this.getMenuElements() }
                    position={ this.props.position }
                    display={ !!this.state.displayMenu } />
-
+               <div style={{postion: 'fixed', top:'1%', right:'1%'}}>
+                   {this.state.event ? this.state.event.pageX : null }
+               </div>
                { this.renderView() }
            </div>
         );